requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
sentence-transformers>=2.2.0
haystack-ai
elasticsearch-haystack
openai>=1.0.0
datasets
